requirements.txt
appnope==0.1.0
boto3==1.1.4
botocore==1.2.7
cffi==1.1.2
characteristic==14.3.0
cryptography==0.9.3
cssselect==0.9.1
dateparser==0.3.0
decorator==4.0.4
docutils==0.12
enum34==1.0.4
facepy==1.0.6
Flask==0.10.1
functools32==3.2.3-2
futures==2.2.0
gnureadline==6.3.3
idna==2.0
ipaddress==1.0.14
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
jsonschema==2.5.1
loginform==1.0
lxml==3.4.4
MarkupSafe==0.23
ndg-httpsclient==0.4.0
numpy==1.9.2
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
pprintpp==0.2.3
ptyprocess==0.5
py==1.4.30
pyasn1==0.1.8
pyasn1-modules==0.0.6
pycparser==2.14
pymongo==3.0.3
PyMySQL==0.6.6
pyOpenSSL==0.15.1
pytest==2.7.2
python-dateutil==2.4.2
PyYAML==3.11
queuelib==1.2.2
requests==2.5.3
rfc3987==1.3.4
scrapely==0.12.0
Scrapy==1.0.1
scrapyd==1.1.0
scrapyd-client==1.0.1
scrapyjs==0.1.1
service-identity==14.0.0
simplegeneric==0.8.1
six==1.9.0
slybot==0.12.1
SQLAlchemy==1.0.8
SQLAlchemy-Utils==0.30.14
TinyUrl==0.1.0
traitlets==4.0.0
Twisted==15.2.1
w3lib==1.11.0
Werkzeug==0.10.4
zope.interface==4.1.2
cov-core==1.15.0
coverage==3.7.1 
pytest-cov==1.8.1 
wheel==0.24.0
